+++ Changelog.md
@@ -34,6 +34,8 @@ v2.38.0 (not published yet)
 - UI: Make Preview a normal window by default
 - UI: Add "Output Mod Direction" to allow decreasing when auto cropping or correcting is applied ([#1285](/../../issues/1285))
 - UI: Fix Import Command Line for checkboxed parameters
+- UI: Add Plugin fulfillment requirements
+    - If a requirement is not fulfilled, the import will be commented out
 - NVEncC: Add new --lookahead-level parameter
 - NVEncC: Add new --tf-level parameter
 - NVEncC: Add new --tune parameter